Set content length on brickftp uploads
Brickftp requires the content length to be set for uploads. In testing I had
used a bytes.Buffer which sets the content length automatically, this does
not happen for generic io.Readers. As per docs:

If body is of type *bytes.Buffer, *bytes.Reader, or *strings.Reader, the
returned request's ContentLength is set to its exact value (instead of
-1)